requirements.txt
# Bot
aiogram>=3.0,<4.0
aiogram==3.10.0

# Framework
Django>=5.0,<6.0
django-ninja>=1.0,<2.0
psycopg[binary]>=3.1
celery[redis]>=5.3

# AI
google-genai
protobuf

# Validation & Config
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv==1.0.0

# Google Sheets (legacy, optional)
gspread-asyncio>=2.0,<3.0
google-auth==2.22.0

# Storage
boto3>=1.34

# Browser automation
playwright>=1.40

# HTTP
aiohttp